requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2   #to extract text from pdf file